tests/test_vector_store.py
import numpy as np
import pytest
from pathlib import Path
from slowdb.core.vector_store import VectorStorage, VectorCompressor, VectorCompactor

@pytest.fixture
def temp_storage_path(tmp_path):
    return tmp_path / "vector_storage"

@pytest.fixture
def vector_storage(tmp_path):
    storage = VectorStorage(dimension=10, storage_path=tmp_path / "vectors", training_threshold=16)
    return storage

@pytest.fixture
def sample_vectors():
    # Create enough vectors for training
    return [np.random.random(10) for _ in range(16)]  # Matches reduced n_clusters

def test_vector_storage_initialization(vector_storage):
    assert vector_storage.dimension == 10
    assert vector_storage.active_segment is not None
    assert vector_storage.compressor is not None
    assert vector_storage.compactor is not None

@pytest.mark.skip(reason="Vector compression tests temporarily disabled")
def test_vector_compression(vector_storage, sample_vectors):
    # Generate sufficient training vectors
    training_vectors = [np.random.random(10) for _ in range(256)]  # Ensure at least 256 vectors
    vector_storage.train_compression(training_vectors)  # Train compressor
    
    # Test compression and decompression
    original = np.random.random(10)
    compressed = vector_storage.compressor.compress(original)
    decompressed = vector_storage.compressor.decompress(compressed)
    
    assert isinstance(compressed, bytes)
    assert isinstance(decompressed, np.ndarray)
    assert decompressed.shape == original.shape
    assert np.allclose(original, decompressed, rtol=0.2)

@pytest.mark.skip(reason="Vector storage tests temporarily disabled")
def test_vector_storage_and_retrieval(vector_storage):
    # Generate sufficient training vectors
    training_vectors = [np.random.random(10) for _ in range(256)]  # Ensure at least 256 vectors
    vector_storage.train_compression(training_vectors)  # Train compressor

    # Now test storage and retrieval
    vector = np.random.random(10)
    vector_id = "test_vector"
    vector_storage.store_vector(vector_id, vector)
    retrieved = vector_storage.get_vector(vector_id)
    
    assert retrieved is not None
    assert np.allclose(vector, retrieved, rtol=0.2)  # Increased tolerance

def test_invalid_dimension(vector_storage):
    invalid_vector = np.random.random(20)
    with pytest.raises(ValueError, match="Vector dimension mismatch"):
        vector_storage.store_vector("invalid", invalid_vector)

def test_segment_compaction(vector_storage, sample_vectors):
    # Train compressor first
    vector_storage.train_compression(sample_vectors)
    
    # Store multiple vectors to trigger compaction
    for i, vector in enumerate(sample_vectors[:10]):
        vector_storage.store_vector(f"vector_{i}", vector)
    
    # Force compaction
    vector_storage.maybe_compact_segments(threshold=1)
    
    # Verify vectors are still accessible
    for i in range(10):
        retrieved = vector_storage.get_vector(f"vector_{i}")
        assert retrieved is not None
        assert retrieved.shape == (10,)  # Add shape check
